Re-introduce in-Rust SSH key setup
The reason for using shell scripts instead was to avoid issues with
gpg-as-ssh-agent, but we can work around that reasonably enough. It's
more likely that most people would have a standard `ssh-agent` running
rather than `gpg-agent`, so let's optimize for them to be able to run
the tests with minimal setup.